requirements.txt
openai-whisper==20240930
numpy==2.0.2
Flask==3.1.0